feat: route instant start through init
